+++ html_report/scripts/build-standalone.mjs
@@ -43,6 +43,8 @@ function localizeFiguresBody(body) {
 function localizeRubricsBody(body) {
   return body
     .replaceAll('href="/"', 'href="./DeepAlign-Bench_HTML汇报版.html"')
+    .replaceAll('href="/data_factory.protocol.yaml"', 'href="./data_factory.protocol.yaml"')
+    .replaceAll('href="/rubric_module_library.yaml"', 'href="./rubric_module_library.yaml"')
     .replaceAll('href="/case.schema.yaml"', 'href="./case.schema.yaml"')
     .replaceAll('href="/rubric_template_registry.yaml"', 'href="./rubric_template_registry.yaml"')
     .replaceAll('href="/rubric_leaf.schema.yaml"', 'href="./rubric_leaf.schema.yaml"')